requirements.txt
Django==1.9
Pillow==3.1.0
wheel==0.24.0
django-markdown-deux==1.0.5
markdown2==2.3.1
django-pagedown==0.1.1
djangorestframework==3.3.3
Markdown==2.6.6
django-filter==0.13.0
django-crispy-forms==1.6.0
